Move compiler driver code from ops into core
