+++ src/cargo/core/mod.rs
@@ -21,6 +21,7 @@ pub mod resolver;
 pub mod summary;
 pub mod shell;
 pub mod registry;
+pub mod compiler;
 mod interning;
 mod package_id_spec;
 mod workspace;
