+++ src/cargo/ops/cargo_compile.rs
@@ -28,8 +28,10 @@ use std::sync::Arc;
 
 use core::{Package, Source, Target};
 use core::{PackageId, PackageIdSpec, Profile, Profiles, TargetKind, Workspace};
+use core::compiler::{BuildConfig, BuildOutput, Compilation, Context, DefaultExecutor, Executor};
+use core::compiler::{Kind, TargetConfig, Unit};
 use core::resolver::{Method, Resolve};
-use ops::{self, BuildOutput, Context, DefaultExecutor, Executor, Kind, Unit};
+use ops;
 use util::config::Config;
 use util::{profile, CargoResult, CargoResultExt};
 
@@ -193,7 +195,7 @@ pub enum CompileFilter {
 pub fn compile<'a>(
     ws: &Workspace<'a>,
     options: &CompileOptions<'a>,
-) -> CargoResult<ops::Compilation<'a>> {
+) -> CargoResult<Compilation<'a>> {
     compile_with_exec(ws, options, Arc::new(DefaultExecutor))
 }
 
@@ -201,7 +203,7 @@ pub fn compile_with_exec<'a>(
     ws: &Workspace<'a>,
     options: &CompileOptions<'a>,
     exec: Arc<Executor>,
-) -> CargoResult<ops::Compilation<'a>> {
+) -> CargoResult<Compilation<'a>> {
     for member in ws.members() {
         for warning in member.manifest().warnings().iter() {
             if warning.is_critical {
@@ -224,7 +226,7 @@ pub fn compile_ws<'a>(
     source: Option<Box<Source + 'a>>,
     options: &CompileOptions<'a>,
     exec: Arc<Executor>,
-) -> CargoResult<ops::Compilation<'a>> {
+) -> CargoResult<Compilation<'a>> {
     let CompileOptions {
         config,
         jobs,
@@ -844,7 +846,7 @@ fn scrape_build_config(
     config: &Config,
     jobs: Option<u32>,
     target: Option<String>,
-) -> CargoResult<ops::BuildConfig> {
+) -> CargoResult<BuildConfig> {
     if jobs.is_some() && config.jobserver_from_env().is_some() {
         config.shell().warn(
             "a `-j` argument was passed to Cargo but Cargo is \
@@ -875,7 +877,7 @@ fn scrape_build_config(
     let jobs = jobs.or(cfg_jobs).unwrap_or(::num_cpus::get() as u32);
     let cfg_target = config.get_string("build.target")?.map(|s| s.val);
     let target = target.or(cfg_target);
-    let mut base = ops::BuildConfig::new(&config.rustc()?.host, &target)?;
+    let mut base = BuildConfig::new(&config.rustc()?.host, &target)?;
     base.jobs = jobs;
     base.host = scrape_target_config(config, &base.host_triple)?;
     base.target = match target.as_ref() {
@@ -885,9 +887,9 @@ fn scrape_build_config(
     Ok(base)
 }
 
-fn scrape_target_config(config: &Config, triple: &str) -> CargoResult<ops::TargetConfig> {
+fn scrape_target_config(config: &Config, triple: &str) -> CargoResult<TargetConfig> {
     let key = format!("target.{}", triple);
-    let mut ret = ops::TargetConfig {
+    let mut ret = TargetConfig {
         ar: config.get_path(&format!("{}.ar", key))?.map(|v| v.val),
         linker: config.get_path(&format!("{}.linker", key))?.map(|v| v.val),
         overrides: HashMap::new(),
