+++ src/cargo/ops/mod.rs
@@ -2,10 +2,6 @@ pub use self::cargo_clean::{clean, CleanOptions};
 pub use self::cargo_compile::{compile, compile_with_exec, compile_ws, CompileOptions};
 pub use self::cargo_compile::{CompileFilter, CompileMode, FilterRule, MessageFormat, Packages};
 pub use self::cargo_read_manifest::{read_package, read_packages};
-pub use self::cargo_rustc::{Compilation, Kind, Unit};
-pub use self::cargo_rustc::{is_bad_artifact_name, Context};
-pub use self::cargo_rustc::{BuildConfig, BuildOutput, TargetConfig};
-pub use self::cargo_rustc::{DefaultExecutor, Executor};
 pub use self::cargo_run::run;
 pub use self::cargo_install::{install, install_list, uninstall};
 pub use self::cargo_new::{init, new, NewOptions, VersionControl};
@@ -38,7 +34,6 @@ mod cargo_package;
 mod cargo_pkgid;
 mod cargo_read_manifest;
 mod cargo_run;
-mod cargo_rustc;
 mod cargo_test;
 mod lockfile;
 mod registry;
